requirements.txt
python-telegram-bot
numpy
openai
sentence_transformers
torch
